+++ src/pages/categoryOptionGroups/form/categoryOptionGroupSchema.ts
@@ -1,25 +1,27 @@
 import { z } from 'zod'
 import { getDefaults, createFormValidate, modelFormSchemas } from '../../../lib'
-import { CategoryOptionGroupSet } from '../../../types/generated'
+import { CategoryOptionGroup } from '../../../types/generated'
 
+/*  Note that this describes what we send to the server, 
+    and not what is stored in the form. */
 const { identifiable, referenceCollection, withAttributeValues } =
     modelFormSchemas
 
-export const categoryOptionGroupSetSchema = identifiable
+export const categoryOptionGroupSchema = identifiable
     .merge(withAttributeValues)
     .extend({
         shortName: z.string().trim(),
         code: z.string().trim().optional(),
         description: z.string().trim().optional(),
         dataDimensionType: z
-            .nativeEnum(CategoryOptionGroupSet.dataDimensionType)
-            .default(CategoryOptionGroupSet.dataDimensionType.DISAGGREGATION),
+            .nativeEnum(CategoryOptionGroup.dataDimensionType)
+            .default(CategoryOptionGroup.dataDimensionType.DISAGGREGATION),
         dataDimension: z.boolean().default(true),
-        categoryOptionGroups: referenceCollection
+        categoryOptions: referenceCollection
             .min(1, 'At least one category option is required')
             .default([]),
     })
 
-export const initialValues = getDefaults(categoryOptionGroupSetSchema)
+export const initialValues = getDefaults(categoryOptionGroupSchema)
 
-export const validate = createFormValidate(categoryOptionGroupSetSchema)
+export const validate = createFormValidate(categoryOptionGroupSchema)